Infer challenge scope if not provided by registry
A registry is expected to provide a scope value when proposing a
challenge. If it fails to do so, it is possible to infer it from the
user request. Currently, Podman and Docker do this because they use the
challenge provided by a "ping" request to fetch a Bearer token, when
possible. This was done in Podman just to allow compatibility with
gcr.io. See [1] for reference.

[1] https://github.com/containers/image/pull/211

Signed-off-by: Athos Ribeiro <athoscr@fedoraproject.org>